pyproject.toml
[project]
name = "mcp-telegram-bot"
version = "0.1.0"
description = "BotBuilder - Expert Telegram Bot Architect that generates and deploys bots automatically"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiogram>=3.0.0",
    "jinja2>=3.1.0",
    "python-dotenv>=1.0.0",
    "aiofiles>=23.0.0",
    "docker>=6.0.0",
    "aiohttp>=3.9.0",
]
